Add 2d proximity

+++ include/game_logic/ai_characters/hider.h
@@ -9,7 +9,7 @@ struct hider
 };
 
 
-struct hider* initialize_hider(int num_chests, int **payoff_matrix);
+struct hider* initialize_hider(int num_chests, double **payoff_matrix);
 void find_hider_strategy(struct hider *hider, int num_chests);
 
 #endif // HIDER_H
